Move supporting files into a subdirectory
Supporting files such as the .xcf's used to create the .gif images for
use in the static part of the site have been moved into a subdirectory
called "supporting" so as not to clutter the main operational
directory.  Supporting files will NOT be copied to the server in the
live system.