Update ie css to make work
This is a bit of a hack, because without IE v 6 available to test I
can't get it right.  However it appears the input box floats to the
top of the div in IE 6 and to the bottom of the div in firefox.

So we move the input container down in IE 6 so it does not overlap
chat list