Fix race condition with new whisper
It is possible for the person you are about to set up a whisper to to
log out at the exact same moment.  This gives a db foreign key error when trying
to create the "participant" table entries for a person who has just
been deleted from the users database.

The trick it to ignore the errors and to also then not create the
whisper box at all