Fix private room exit problem
When exiting a private room, the software tries to re-activate your
whisper box by unhiding it.  However, it might have been removed
already by the other whisperers exiting their whisper.  If this is the
case we mustn't do this because it leaves you locked in your private
room unable to get out.