Fixup problem whereby when a user is logged out because of a comms timeout his pipe was not deleted.
When a users comms fail, he does not go through the normal logout
process which deletes his communications pipe.  So when he connects
again, he will be immediately logged out in order to delete it.

This commit changes it so that when you detect a timeout, you delete
the pipe of the user so when he restarts he can restart properly.

In this case you also have to cater for the fact that in the
Javascript, if you see a LT message (logout with timeout) you try and
log in again.  In this case you need to recreate the fifo.

There was a 5 second delay in logging out - this has been changed to
be only 2 seconds.