Seems now to be working under load
A number of changes have been made - perhaps the most significant is
that the writing of the log message into the database and sending it
to all participants down the pipe has been combined into one file.  It
has not yet been transactionised but it doesn't appear necessary

Read has been altered so that if a new message appears whilst the read
request is not outstanding (in the small gap between completing the
read and puting up the next request) it queries the database for the
missing entries.

This version includes a test program in the test directory - it will
need to be modified as to where the base url is.