Snapshot in the building of a separate server implementation.
It is almost working, but I am now making some crucial changes and I
don't want to loose anything.  So this snapshot just means I can go
back easily if I need to

NOTE:  The snapshot as it stands does not work - because client
expects to start the server with the data dir as a parameter, but the
server does not yet read it.  That will be the next modification