Internet Explorer fix - it needs input elements (username and password) to be given full element properties
The two input fields were being used to add error class to.  This
didn't work because the elements had not been given the full element
properties.  It was necessary to wrap them in a document.id in order
to ensure they had the removeClass and addClass methods.

I am not sure why this DID work before with other browsers.