Fixed rubbish entries at the end of some DO arrays
It turns out that many of the entries in the DO arrays had a comma
after the last entry - in effect causing a dummy additional entry into
the array with a "fail" type.  For reasons that I don't quite know,
Internet Explorer balked on these (where as the other browsers didn't
seem to notice them).  This correction removes them.

In addition, each NS context seems to hold an additional 'name' field
that wasn't being used, so I removed it.