Rework of authentication to provide hooks for remote authentication
Changed the authetication process so that it first tries to
authenticate with no username before putting up the input form.  This
allows external authentication modules to hook in here and respond
with credentials it may have retrieved via other routes.  For instance
Melinda's Backups looks up the user in the forum database based on a
cookie that the user has.

A special response - I have logged in as user uid=0 allows us to
detect that we are really being asked to put up the login form and ask
for a user name and password.

At the same time, all the authentication stuff has been carved of into
a separate loadable .js file mbcauth.js so that we can later in the
development process produce a compressed version (we may also bundle
all the cipher files together so as have a single loadable
authentication suite).