Some updates to fix url issues with chat, plus add blind support for whispering
Blind people using jaws can see input buttons easily.  So if we have a
blind users, we have removed the exit div and replaced it with a
button.  In addition we change the online list to be a list of buttons
rather than divs which contain the user.  This enables the easy
selection of one of these buttons to start a whisper.

Also they cannot use "Enter" to send a form, as its used by Jaws to
change mode.  So we are using cntl s instead.  So there is no need for
the send buttons in the main chat area and in the whisper boxes.  To
avoid clutter for them these have been removed as well.