Relax security controls and remove the time based element of passwords.
This has proven necessary because of the variability of clients out
there.  It seems that there clocks are not well enough synchronised to
the real time to make this a practical possibility.  So instead we are
reverting to the "REMOTE_KEY" concept in which the md5 or the remote
key is used as the password.  This is slightly less secure, but does
still require the potential hacker to find out what it is