Fix some issues in chatter, where it looked up non existant records
Chatter was using querySingle to look up data and assuming that if it
didn't find it the query would return false.  It doesn't it returns
and empty array, so I have changed it to check for set variables
related to the primary key.  If its there we have a record - if not we
didn't get one.